pyproject.toml
[tool.poetry]
name = "transaction-analytic"
version = "0.1.0"
description = ""
authors = ["<Mariya Polshkova> <<polshkovamamariya@gmail.com>>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.1.3"
xlrd = "^2.0.1"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
openpyxl = "^3.1.2"
pandas-stubs = "^2.1.1.230928"
types-requests = "^2.31.0.10"


[tool.poetry.group.lint.dependencies]
mypy = "^1.7.1"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"



[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py

)
'''

[tool.isort]
line_length = 119

[tool.mypy]
no_implicit_optional = true
exclude = ['venv', 'tests']


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


